refactored views.py to function based files
